ndrop - move code from ndrop.hpp to nndrop_od.hppp. Rename ndrop_od.hpp.
