+++ src/mam4xx/mam4.hpp
@@ -22,7 +22,6 @@
 #include <mam4xx/hetfrz.hpp>
 #include <mam4xx/mam4_types.hpp>
 #include <mam4xx/ndrop.hpp>
-#include <mam4xx/ndrop_od.hpp>
 #include <mam4xx/nucleate_ice.hpp>
 #include <mam4xx/nucleation.hpp>
 #include <mam4xx/rename.hpp>
