+++ src/validation/ndrop/loadaer.cpp
@@ -87,11 +87,11 @@ void loadaer(Ensemble *ensemble) {
 
     Real aten = zero;
 
-    ndrop_od::ndrop_int(exp45logsig, alogsig, aten,
+    ndrop::ndrop_int(exp45logsig, alogsig, aten,
                         num2vol_ratio_min_nmodes,  // voltonumbhi_amode
                         num2vol_ratio_max_nmodes); // voltonumblo_amode
 
-    ndrop_od::loadaer(
+    ndrop::loadaer(
         state_q.data(), nspec_amode, air_density, phase, lspectype_amode,
         specdens_amode, spechygro, lmassptr_amode, num2vol_ratio_min_nmodes,
         num2vol_ratio_max_nmodes, numptr_amode, qcldbrn, qcldbrn1d_num.data(),
